Start moving logic for creation from CartServiceImpl to CartJpaRepository class
