+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/CartRepository.java
@@ -1,13 +1,9 @@
 package com.avdo.spring.app.repository;
 
-import com.avdo.spring.app.entity.CartEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
+import com.avdo.spring.app.service.domain.model.Cart;
 
-import java.util.Optional;
-
-@Repository
 public interface CartRepository {
 
-    Optional<CartEntity> findByUserEntityId(Long userId);
+    Cart findByUserEntityId(Long userId);
+    Cart createCart();
 }
