+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/impl/CartJpaRepository.java
@@ -1,6 +1,44 @@
 package com.avdo.spring.app.repository.impl;
 
+import com.avdo.spring.app.entity.CartEntity;
+import com.avdo.spring.app.entity.UserEntity;
 import com.avdo.spring.app.repository.CartRepository;
+import com.avdo.spring.app.repository.crud.CrudCartRepository;
+import com.avdo.spring.app.service.domain.model.Cart;
+import com.avdo.spring.app.service.domain.model.User;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Repository;
 
+import java.sql.Date;
+import java.time.LocalDate;
+
+@Repository
 public class CartJpaRepository implements CartRepository {
+
+    private final CrudCartRepository crudCartRepository;
+
+    @Autowired
+    public CartJpaRepository(CrudCartRepository cartRepository) {
+        this.crudCartRepository = cartRepository;
+    }
+
+    @Override
+    public Cart findByUserEntityId(Long id) {
+
+    }
+
+    @Override
+    public Cart createCart() {
+        UserEntity userEntity = extractUserFromToken();
+        User user = userService.findById(id);
+        CartEntity cartEntity = new CartEntity();
+        cartEntity.setUserEntity(UserEntity.fromUser(user));
+        cartEntity.setDateCreated(Date.valueOf(LocalDate.now()));
+        return crudCartRepository.save(cartEntity).toDomainModel();
+    }
+
+    private UserEntity extractUserFromToken() {
+        return (UserEntity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    }
 }
